Revert "개인 프로필 조회 기능"

+++ src/main/java/clofi/runningplanet/member/dto/response/ProfileResponse.java
@@ -9,17 +9,17 @@ public record ProfileResponse(
 
 	Gender gender,
 
-	Integer age,
+	int age,
 
 	String profileImg,
 
-	Integer runScore,
+	int runScore,
 
 	AvgPace avgPace,
 
-	Integer avgDistance,
+	double avgDistance,
 
-	int totalDistance,
+	double totalDistance,
 
 	String myCrew
 ) {
